inference_docker/requirements.txt
torch==2.4.1
torchvision==0.19.1
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
pillow